linux-kthread: swap ptid lwp and tid usage to match GDB remote.
Swap the usage of ptid to match GDB remote which has the
following layout

* GDB Remote uses LWP to store the effective cpu core
 *  ptid.pid = Inferior PID
 *  ptid.lwp = CPU Core
 *  ptid.tid = 0

Formally lkd used the following layout:
 * STMC Debug remote uses TID to store the core.
 *  ptid.pid = Inferior PID
 *  ptid.lwp = Process ID
 *  ptid.tid = CPU Core (-1 for not running)

linux-kthread now also uses -1 to symbolise 'not running' for lwp.

The only downside to this change is when new threads are created
gdb automatically prints the lwp. In the old layout this matched
with the Linux process ID, and now normally prints -1 (to indicate
the thread is not running).

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>