Add Linux kernel thread runtime support.
This patch implements a new linux-kthread target_ops stratum which
supports the Linux kernel thread runtime. It allows anything
that the Linux kernel has created a task_struct for to be represented
as a GDB thread object. This allows a user using GDB to debug the
Linux kernel to see all the sleeping threads in the system rather than
just physical CPU's (as is the case currently) and then use the GDB
contextual commands such as 'thread' to easily switch between all the
threads in the system, inspect data structures and get backtraces etc.

e.g.
(gdb) info threads
  Id   Target Id         Frame
* 1    [swapper/0] (pid: 0 tgid: 0 <C0>) cpu_v7_do_idle ()
    at /sources/linux/arch/arm/mm/proc-v7.S:75
  2    init (pid: 1 tgid: 1) context_switch (cookie=...,
       next=<optimized out>, prev=<optimized out>, rq=<optimized out>)
    at /sources/linux/kernel/sched/core.c:2902
  3    [kthreadd] (pid: 2 tgid: 2) context_switch (cookie=...,
       next=<optimized out>, prev=<optimized out>,
    rq=<optimized out>) at /sources/linux/kernel/sched/core.c:2902
<snip>
  90   getty (pid: 1584 tgid: 1584) context_switch (cookie=...,
       next=<optimized out>, prev=<optimized out>,
    rq=<optimized out>) at /sources/linux/kernel/sched/core.c:2902
  91   udevd (pid: 1586 tgid: 1586) context_switch (cookie=...,
       next=<optimized out>, prev=<optimized out>,
    rq=<optimized out>) at /sources/linux/kernel/sched/core.c:2902

linux-kthread.ch is split between the linux-kthread target_ops stratum
methods themselves which are fairly self explanatory, helper functions
which parse kernel data structures such as CPU runqueue for idle and curr
tasks and a series of low level helper functions and macros which make
obtaining symbol information and calculating struct field offsets
of the various Linux kernel data structures much easier.

Architecture support is implemented in the <arch>-linux-kthread files,
and is currently limited to ARM, although more architecture support
is expected to follow. Adding new architecture support is fairly
trivial and mainly centres around populating the register cache
of a sleeping thread from what the kernel saved on the stack when
the thread was de-scheduled.

    gdb/ChangeLog (Peter Griffin):

        * linux-kthread.c, linux-kthread.h, arm-linux-kthread.c,
          arm-linux-kthread.h: New files.
        * configure.tgt: Add linux-kthread.o and arm-linux-kthread.o to
          gdb_target_obs.
        * Makefile.in: Add arm-linux-kthread.o and linux-kthread.o to
          ALL_TARGET_OBS.
        * gdbarch.sh (linux_kthread_arch_ops): New
        * gdbarch.c, gdbarch.h: Re-generated.
        * arm-tdep.c (arm_gdbarch_init): Call register_arm_linux_kthread_ops.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>